# Conflicts:
#	src/constants.py
#	src/main.py
#	src/player/actions.py
#	src/player/player.py
#	src/player/player_thinker.py
#	src/player/playerstrategy.py
#	src/uppaal/models/staminamodel/staminamodell11.xml
#	src/uppaal/models/staminamodel/staminamodell6.xml
#	src/uppaal/models/staminamodel/staminamodell7.xml
#	src/uppaal/models/staminamodel/staminamodell8.xml
#	src/uppaal/models/staminamodel/staminamodell9.xml
#	src/uppaal/strategy.py
#	src/uppaal/uppaal_model.py

+++ src/soccer_sim.py
@@ -12,7 +12,7 @@ from statisticsmodule import log_parser
 
 class SoccerSim(threading.Thread):
     def __init__(self, team_names: [str], num_players: int, trainer_mode: bool, coaches_enabled: bool, udp_player: int,
-                 udp_trainer: int, udp_coach: int, udp_ip: str) -> None:
+                 udp_trainer: int, udp_coach: int, udp_ip: str, enable_monitor: bool) -> None:
         super().__init__()
         self.team_names = team_names
         self.num_players = num_players
@@ -32,6 +32,7 @@ class SoccerSim(threading.Thread):
         # References to the simulator
         self.soccer_sim = None
         self.soccer_monitor = None
+        self.enable_monitor = enable_monitor
 
         self.has_init_clients = False
 
@@ -49,7 +50,8 @@ class SoccerSim(threading.Thread):
 
         # Use soccerwindow2: exec soccerwindow2 --kill-server --geometry=1440x900 --gradient 1 --field-grass-type lines
         # Use regular monitor: exec rcssmonitor --show-status-bar 1 --show-kick-accel-area 1 --show-catch-area 1 --geometry=1280x800
-        self.soccer_monitor = subprocess.Popen(["exec rcssmonitor --show-status-bar 1 --show-kick-accel-area 1 --show-catch-area 1 --geometry=1280x800"], shell=True)
+        if self.enable_monitor:
+            self.soccer_monitor = subprocess.Popen(["exec rcssmonitor --show-status-bar 1 --show-kick-accel-area 1 --show-catch-area 1 --geometry=1280x800"], shell=True)
 
 
     def run(self) -> None:
@@ -99,8 +101,9 @@ class SoccerSim(threading.Thread):
         if self.trainer_mode:
             self.trainer.stop()
             self.trainer.join()
-        self.soccer_monitor.send_signal(signal.SIGINT)
-        self.soccer_monitor.wait(3)
+        if self.soccer_monitor is not None:
+            self.soccer_monitor.send_signal(signal.SIGINT)
+            self.soccer_monitor.wait(3)
         self.soccer_sim.send_signal(signal.SIGINT)
         self.soccer_sim.wait(3)
         log_parser.parse_logs()
